feat: generalise `macro.Call` instruction (#1238)
* support call arguments as expressions

This adds support for calling functions wtih expressions as arguments.
There are numerous benefits of this, such as the ability to use
constants in argument positions, and the ability to have different
bitwidths, etc.

* update testcase for generalised macro.Call

This updates various test cases to make use of the improvements to the
macro.Call instruction.  In some cases, this allows us to eliminate
colunms that were previously required only to work around the
limitations of the instruction.