fix: inspector incorrectly displaying register limbs (#1245)
* support TraceView

TraceView provides the high-level abstraction for representing a view of
a trace.  Specifically, the view could be as part of a report for the
check command, or it could what is shown in the inspector, etc.  Thus,
TraceView provides a general purpose abstraction which consolidates a
number of previously separate implementations.

* support windowing

The idea of windowing is to further break up the ModuleView concept to
try and improve the structure.