make constraints split with RegisterAllocator
This completes the loop in terms of a vanishing constraint being turned
into an agnostic.Equation and then back into a constraint.  In
particular, it plumbs through RegisterAllocator instead of LimbsMap.
But this has the negative consequence that lookup splitting no longer
works, since splitting a lookup could introduce columns into the source
and target modules.