dramatically simplified core algorithm
The simplifications come from separating the issue of determining the
chunk points from that of dividing up the polynomials.  This separation
of concerns allows us to use a very naive chunk point calculator for
now, and focus on having something which works.

Currently this uses a naive chunking algorithm, which potentially might
not terminate in general.  However, it is sufficient for now.

This adds support for carry registers to constraints, but does not yet
include assignments for those carry registers.

When a chunk is encountered which remains too large for the underlying
bandwidth, an error is flagged.