feat: support HIR representation (#1255)
This reintroduces an HIR representation which sits between Corset and
MIR.  The HIR representation is, for the most part, exactly the same as
the old MIR representation.  Thus, it is the MIR representation which
has significant changed as a result of this.  Specifically, it is now a
much simpler representation which is closer to the AIR level.  For
example, most constraint forms operate over register accesses only,
rather than arbitrary terms.  This means the lowering from HIR => MIR
requires expanding terms as necessary into columns.

At this stage, there remain some loose ends to be fixed.  For example,
handling of lookups with differing numbers of limbs is not support
correctly.  Likewise, there are optimisations possible for source-level
vector accesses.