switch over more passing tests
Currently, we have around 40% of the Corset valid tests are passing for
the four different fields.

This also simplifies the existing tests by removing occurences of the
"vanishes" pure function which was previously defined in many files.

This also adds "logical equivalences" to a small number of tests.  That
is, where logical operators are used instead of multiplication, etc.