separate RegisterAllocator and RegisterLimbsMap
This separates these two concepts apart, as their combination was
actually confusing.  In particular, the issue was that in both
agnostic.assignment, and agnostic.equation we only need a single
register map (i.e. because splitting is done prior to these algorithms
being run).  Thus, having a register limbs map is then confusing because
we have to remember to distinguish between registers and limbs.