feat: support register splitting for MIR constraints (#1263)
* remove inequality

Whilst inequality's have been supported in some limited scenarios, they
are not actually used in a real-world setting.  Therefore, to simplify
the problem of splitting constraints they have been removed to reduced
the number of constraints forms to consider.

* make constraints split with RegisterAllocator

This completes the loop in terms of a vanishing constraint being turned
into an agnostic.Equation and then back into a constraint.  In
particular, it plumbs through RegisterAllocator instead of LimbsMap.
But this has the negative consequence that lookup splitting no longer
works, since splitting a lookup could introduce columns into the source
and target modules.

* dramatically simplify core algorithm

The simplifications come from separating the issue of determining the
chunk points from that of dividing up the polynomials.  This separation
of concerns allows us to use a very naive chunk point calculator for
now, and focus on having something which works.

Currently this uses a naive chunking algorithm, which potentially might
not terminate in general.  However, it is sufficient for now.

This adds support for carry registers to constraints, but does not yet
include assignments for those carry registers.

When a chunk is encountered which remains too large for the underlying
bandwidth, an error is flagged.

* separate static and relative polynomial

The purpose of this is to allow relative polynomials --- that is
polynomials where the identifier is not just a register identifier, but
a pair of register identifer and shift.  As such, we also now
distinguish a "relative identifier" from a static register identifier.

* switch over more passing tests

Currently, we have around 40% of the Corset valid tests are passing for
the four different fields.

This also simplifies the existing tests by removing occurences of the
"vanishes" pure function which was previously defined in many files.

This also adds "logical equivalences" to a small number of tests.  That
is, where logical operators are used instead of multiplication, etc.

* separate RegisterAllocator and RegisterLimbsMap

This separates these two concepts apart, as their combination was
actually confusing.  In particular, the issue was that in both
agnostic.assignment, and agnostic.equation we only need a single
register map (i.e. because splitting is done prior to these algorithms
being run).  Thus, having a register limbs map is then confusing because
we have to remember to distinguish between registers and limbs.

* initial non-linear splitting algorithm in place

This puts in place an initial version of the non-linear register
splitting algorithm.  This does not, at the current time, actually fill
the introduced limbs.  However, this should be relatively easy to add.

* separate out register package

This separates "register stuff" into a dedicated register package.

* move agnosticity stuff into agnostic package

* relocate ir terms into ir/term

* support polyfil filler

The polyfil filler is essentially a computation for evaluating
polynomials in a word.BigEndian environment and optionally
right-shifting them.

* support limb filling

* add extensive non-linear test

This adds a straight forward non-linear test case for the register
splitting algorithm with around 5000 valid/invalid tests (each).

This also increases the number of tests which are running over three fields.

* support matrix testing per field

This allows testing in the CI pipeline to be done on a per field basis,
meaning that we can have separate runners for the same tests on
different fields.

* downsample asm bench tests

This reduces the number of individual tests for the assembly benchmarks.
Whilst this is less than ideal, we also need to manage the cost of
running the CI pipeline.