feat: improve performance of the trace command (#1270)
* initial update for trace command

This puts through the initial update for the trace command which was
pretty straightforward.  It remains to fix up the various features it
provides, but the basic structure is in place.

* module view is operational again

* column view is operational again

* trace summary operational again

At this stage, there remains a significant discrepency between the
numbers produced now and in the past.  Specifically, I believe the
numbers now do not include any columns generated as a result of trace
expansion.

* fix lints