fix: split interleaving constraints assignments (#1282)
* support splitting interleaving constraints / assignments

This adds support for splitting interleaving constraints.  Specifically,
this is achieved by representing interleaving constraints at the MIR
level using VectorAccess expressions rather than RegisterAccess
expressions. 

* refactor notion of a module name

This refactors the notion of a module's name from a simple string to a
pair of string and length multiplier.  This allows more fine-grained
examination of what exactly a module is.  Specifically, it should help
us to distinguish columns in modules with the same name, but different
multipliers.

* update trace command to manage interleaved modules

This updates the trace command to properly show interleaved modules.

* fix for register splitting

There was a preexisting bug related to register splitting, whereby
computed columns which were being split were handled incorrectly.

* vectorise native interleave / filter computations

This makes use of a new primitive to simplify the process of vectorizing
native computations (and potentially other things).  This is an "array
vector" with a corresponding "mutable array vector".  Thus, we have an
array of values which is split into a vector of limbs.