feat: support register splitting for permutations (#1296)
* update tests to include small fields

* minor

* split permutations and sorted constraints

* support range constraints

This adds support for range constraints, though it is incomplete at the
moment it seems.

* mark tests which are passing / failing

This updates the set of tests known to either pass or fail.  At this
stage, there remain some outstanding issues with e.g. strict sorted
constraints.  For StrictSorted_01 this appears to be related to values
which are too large to fit.

* minor fix for subdivideRange

There was a minor discrepency in the bitwidth calculation.

* minor tweaks

* updating agnostic and bench tests

* report error when value exceeds available limbs

The tool now reports an error when a trace value overflows the allocated
limbs for a given register.

* working through classifying corset_value failures

* more tweaks to tests

* fix: common register width calculation

This calculation was not correct under certain circumstances (e.g. when
the average width was already a power of 2).

* more fixes for tests

* minor fix for stp benchmark

The way in which registers are calculated has slightly changed the
register assignment algorithm.

* minor error fix