feat: change TRM module to recognize secp256r1 prc address (#1281)
* feat: change TRM module to recognize secp256r1 prc address

The TRM module is to trim fields and also to recognize precompile addresses. In Osaka, a new precompile is added at address 0x100. TRM module now recognizes this address as a precompile. 
From London to Prague, the number of contiguous precompiles increased to 0x11. This constant is also updated along with the tests expected results.