feat: support restricted register accesses
This puts in place support for "restricted" or "narrow" register
accesses.  This essentially means that register accesses support an
internal "bitwidth" which determines how much of the enclosing register
is actually read.  For example, we might access only the first u4
component of a u8 register.  In effect, this allows for casts at the
register access level.