feat: enable debugging constraints which explode under small fields (#1307)
* working on VariableSplitter

The purpose of this is to enable a one-pass splitting of variables
within a polynomial to avoid exponential blow up in terms of
performance.

* improved variable splitting algorithm

This previosly had exponential behaviour, but is now running much more
efficiently.

* improve polynomial subdivision algorithm

This was previously struggling at times with performance on large
polynomials.  After some profiling and optimisation work, it is now
much improved.

* support "--exploding-multiplier" threshold

This adds support for a cli option to change the threshold above which a
constraint is considered to be "exploding".  This is really only for
debugging purposes and, eventually, should be removed.

* fix Monomial.Shr()

This was incorrectly handling the negative case.  In fact, negative
shifts must be support for zkasm.