extend register.RelativeId => register.AccessId
This extends the original notion of a "relative register id" to that a
"register access id".  Specifically, the latter includes a relative
shift (as before) along with a "bitwidth window".  The latter can be
viewed as a form of cast which acts directly on a register.  For
example, we could access only the first 8bits of a given u16 register.