feat: support restricted register accesses (#1308)
* feat: support restricted register accesses

This puts in place support for "restricted" or "narrow" register
accesses.  This essentially means that register accesses support an
internal "bitwidth" which determines how much of the enclosing register
is actually read.  For example, we might access only the first u4
component of a u8 register.  In effect, this allows for casts at the
register access level.

* extend register.RelativeId => register.AccessId

This extends the original notion of a "relative register id" to that a
"register access id".  Specifically, the latter includes a relative
shift (as before) along with a "bitwidth window".  The latter can be
viewed as a form of cast which acts directly on a register.  For
example, we could access only the first 8bits of a given u16 register.

* fix RegisterAccess encoding / decoding

* refactor RegisterAccess to distinguish the mask

This refactors RegisterAccess and ColumnAccessId to distinguish the
declared bitwidth from the "mask".  That is, the portion of the
underlying column actually used.

* get precise range analysis for dynamic polynomials

This improves the range analysis for dynamic polynomials.  That is,
polynomials which support masks.

* tweak lookup_16 tests

This test had some (now invalid) expanded tests.  They are invalid
because of a minor optimisation around expanding terms into columns.

* minor bug fixes