fix: signed selector in lookup (#1312)
This updates the way bitwidths are calculated when expanding a term into
a column, which is part of the lowering process from HIR => MIR.  They
key benefit of doing it like this is that, at the HIR level, we still
have a true notion of cast.

The signature of ValueRange() no longer requires a register.Map
parameter because the necessary information is already embedded within
register accesses directly.

This adds a new lookup tests along with a good range of trace inputs.

Since range constraints have their argument lowered into a column, we
actually need to check the value is unsigned.  Otherwise, there is a
problem under a small field interpretation where negative values have no
meaning.