fix: signed selector in lookup (#1312)
* add lookup example

* fix: calculate bitwidths using HIR terms

This updates the way bitwidths are calculated when expanding a term into
a column, which is part of the lowering process from HIR => MIR.  They
key benefit of doing it like this is that, at the HIR level, we still
have a true notion of cast.

* update signature of ValueRange()

The signature of ValueRange() no longer requires a register.Map
parameter because the necessary information is already embedded within
register accesses directly.

* fix lookup test

This fixes the new lookup test, which was previously missing a lot of
tests.

* support sign check for range constraint

Since range constraints have their argument lowered into a column, we
actually need to check the value is unsigned.  Otherwise, there is a
problem under a small field interpretation where negative values have no
meaning.