improve subdivision mechanism
This reworks the subdivision mechanism to consolidate as much as
possible within the MIR level.  This makes sense, since that is the
point a which subdivision arises.  Furthermore, doing it in a concrete
place allows to overcome various challenges, such as managing temporary
registers added across all modules.  For example, a lookup can end up
adding temporary padding registers to all modules involved in the
lookup.