generate assignments for zero registers
This now adds an assignment to fill a zero register.  Eventually, this
should be optimised since its actually not very optimal at all at the
moment.  We need an assignment specifically dedicated towards constant
registers.  This would then construct an appropriate MutArray, roughly
along the lines of what is already done with the ZeroArray.