fix: computed register requires at least one limb (#1313)
* fix: signed selector in lookup (#1312)

This updates the way bitwidths are calculated when expanding a term into
a column, which is part of the lowering process from HIR => MIR.  They
key benefit of doing it like this is that, at the HIR level, we still
have a true notion of cast.

The signature of ValueRange() no longer requires a register.Map
parameter because the necessary information is already embedded within
register accesses directly.

This adds a new lookup tests along with a good range of trace inputs.

Since range constraints have their argument lowered into a column, we
actually need to check the value is unsigned.  Otherwise, there is a
problem under a small field interpretation where negative values have no
meaning.

* distinguish lisp numbers from symbols

The purpose of this is to make quoting work properly.  This then allows
me to distinguish registers for constant values from constant values,
etc.

* improve subdivision mechanism

This reworks the subdivision mechanism to consolidate as much as
possible within the MIR level.  This makes sense, since that is the
point a which subdivision arises.  Furthermore, doing it in a concrete
place allows to overcome various challenges, such as managing temporary
registers added across all modules.  For example, a lookup can end up
adding temporary padding registers to all modules involved in the
lookup.

* generate assignments for zero registers

This now adds an assignment to fill a zero register.  Eventually, this
should be optimised since its actually not very optimal at all at the
moment.  We need an assignment specifically dedicated towards constant
registers.  This would then construct an appropriate MutArray, roughly
along the lines of what is already done with the ZeroArray.