fix: use global handle checking
Previously, the algorithm for checking that two constraints do not have
the same handle (in the same module) only worked when the constraints
were defined within the same module block.  However, if they were
defined in different blocks (e.g. in different files) the duplicate
handle was not reported as an error.

To resolve this, a global map of all handles is now maintained which is
shared across all source files and module blocks.