feat: splitting property assertions (#1321)
* enable previously broken tests

This enables several tests which were previously broken due to the lack
of splitting for property assertions.

* support splitting property assertions

Splitting property assertions is more challenging than expected since we
want to maintain the notion that a property is an execution computation,
rather than a term which can be reduced to an arithmetic expression.

* simplify Assertion type and tidy up

This simplifies the assertion type to take only a single generic
parameter F.