feat: function call declaration (#1329)
This adds support for parsing defcall declarations, along with an
appropriate AST node.

This adds support for the resolution, processing and typing stages for
the DefCall construct.  An initial implementation of translation is also
included, but this does not yet actually translate the call.

This adds support for checking arguments and returns.  In particular,
check that there are enough arguments and returns, and also that their
bitwidths follow the expected subtyping pattern.

Condition function calls are supported where a (logical) selector is 
provided.  Function calls are encoded / decoded at the HIR level.