feat: cherry pick defcall into v11 (#1331)
This adds support for parsing defcall declarations, along with an
appropriate AST node.  The syntax supports source selectors, and
updates encoding / decoding as necessary.

This adds support for the resolution, processing and typing stages for
the DefCall construct.  An initial implementation of translation is also
included, but this does not yet actually translate the call.

This updates the trace propagation algorithm to recognise function
call's.  These can then trigger calls into assembly functions.  One
problem is that these are evaluated in the base field, meaning they
can overflow (in principle).  To address this, a check is used to protect
against unexpected overflow during trace propagation of function calls.