feat: optimise branch table reasoning (#1336)
This puts through several improvements to the logical reasoning
framework.  For example, additional rewrites for cases involving unit
propagation.  This unlocks several benchmarks which previously could
not be compiled on KOALABEAR_16 in practical time.  There are still
some missing rules, so this could be further improved in time.

Finally, this adds a number of tests to the logical reasoning framework.