compile out ternary operator
The ternary operator could not be split easily to a small field.
Furthermore, since branch table optimisation has improved the generated
constraints we can (for now at least) forgo the benefits of the ternary
operator.  Thus, it is now compiled out into skip statements during the
transition from macro to micro instructions.

In addition, this delievers an upgrade to the ternary operator, as it
now supports arbitrary true / false branches.