ensure safe evaluation for conditional lookups
This puts in place a mechanism to prevent cast failures for conditional
lookups.  Specifically, the problematic situation was where a value
which caused a cast failure existed on a row where the selector was 0.
In such case, the term did not need to be evaluated and this was causing
problems.