feat: cli flag to prevent trace propagation failure (#1395)
* support tracing despite propagation failure

This updates the trace and inspect commands to work even when a trace
propagation failure occurs.

* support private modules in inspector

The inspector had a problem in that it could not show private modules;
furthermore, in the event there we no public modules then it crashed.
This is now resolved by providing a toggle to show private modules.