ensure target registers always stored in reverse (#1396)
A problem with the ternary instruction was that it stored the target
registers in the declaration order.  In contrast, the assignment
operator stores them the reverse declared order.  However, the call
instruction also stored them in the declared order!

Therefore, this now ensures all three forms store their target registers
in the reverse declared order.