Feat/blake2f (#1390)
Add blake2f implementation which utilises bit_ror for the rotate right operator.
Since zkasm lacks support for static reference tables, the sigma table is 
implemented inline as a function --- which is far from ideal.  The intention is
that this will be replaced once zkasm supports static tables.  Furthermore,
support for static arrays within a function would make the overall look of the
functions neater.

---------

Co-authored-by: DavePearce <dave01001110@gmail.com>