feat: add sampled mul tests (#1380)
* add sampled mul tests

This reduces the number of test inputs for the mul benchmark in order
that it can complete in some reasonable time

* resample asm bench tests / tweaks

This resample the bench tests for euc, exp and wcp.  This also tweaks
the settings for GOMAXPROCS and GOMEMLIMIT (which appears to help
considerably), and also prevents parallel testing on the larger tests
(i.e. because go-corset is already doing stuff in parallel for them).