feat: support division instruction (#1398)
* add macro division instruction

This adds support for the macro division instruction, including parsing.
At this stage, it remains to introduce a corresponding micro
instruction, etc.

* refactor ite (macro) instruction

This refactors the ite instruction to follow the layout used in the new
division instruction.  This is neater, and makes life a bit easier.

* refactor ifgoto (macro) instruction

This refactors the ifgoto instruction to follow the layout used in the
division instruction.  This is neater, and makes life a bit easier.

* add division micro-code

This adds a new micro-code for division, and also refactors the existing
skip code to use an expression form.

* initial translation of division micro-code

This puts in place an initial translation of the division micro-code.  A
key outstanding issue is that it does not yet enforce the requirement
that remainder < divisor must always hold.  A number of tests have been
added, of which a chunk are failing for this reason.

* support witness for division instruction

This adds the notion of an explicit witness for the division
instruction.  In principle, this could eventually be inferred.  For now,
it is required to be given explicitly.

* add reject tests for division

This adds a range of reject tests for division which help to ensure that
it is translated soundly.