refactor division instruction to compute & check
This refactors the way the division instruction to employ the "compute &
check" paradigm.  This means it first computes values and then check
they are, in fact, the right values.  This makes sense from the
perspective of register splitting, since splitting multiplication is:
(i) easier; (2) already implemented.  The downside, however, is that it
somewhat breaks the general approach taken for all other instructions.