update use of division instruction
This updates the manner in which the division instruction is used to
confirm to the current requirements (i.e. to provide a witness, a sum
and a witness sum).  These should be relaxed in the near future,
however.

The key purpose of this update is to utilise the fact that the updated
division instruction now supports register splitting and, hence, we can
run this benchmark under KOALABEAR_16.