preliminary support for unsafe casts
This adds preliminary support for unsafe casts.  They are unsafe in the
sense that the underlying system can be unsound if the cast is violated.
Outstanding issues: (1) runtime cast checking? (2) should cases be safe?

My feeling is that they should be safe, with the proviso that they can
be marked as unsafe by an internal analysis.  That is, when the tool can
prove they are always safe, no additional machinery is required to
enforce that.