fix: panic on missing sign bit
This puts in place a fix to resolve the potential panic around missing
sign bits.  The problem arose during register splitting when the last
chunk of a signed assignment contained only the sign bit.  In such case,
we end up adding a "borrow bit" to connect the two chunks when, in fact,
the original sign bit served the exact purpose.  Therefore, the
algorithm now considers this specific case and will include the sign bit
in the final chunk when, otherwise, it would be in an chunk on its own.