feat: add mod benchmark (#1362)
* add preliminary mod implementation

This adds a preliminary implementation of the MOD module.  The main
outstanding issue is the fact that there is no support at this time for
division.  Realistically, this is needed since otherwise the
implementation is incredibly inefficient.

* exploit division instruction

This updates the benchmark to use the new division instruction, rather
than implementing division using a wasteful loop.

* fix SMOD semantics

The semantics of SMOD are somewhat unusual in that only the sign of the
numerator is taken into consideration as to determine when the result is
negative.