feat: add implementation for oob (#1407)
* add implementation for oob

This adds an implementation of the Out-Of-Bounds (OOB) module which
breaks into three components; checking simple instructions (e.g.
jump/call/create); checking precompiles (except modexp); checking the
modexp precompile.

The implementation aims retain identical behaviour to the original and,
hence, suffers some inefficiency and lack of comprehensibility as a
result.  The intention would be in the future to refactor this module.
Further optimisations are also likely to be beneficial.

At this stage, testing this module runs relatively slowly compared with
others and that remains to be investigated.  Furthermore, it only
supports BLS12_377 due to the lack of a splitting algorithm for the
division instruction.

Finally, test cases for all instructions are included in the default
accepts file.  However, there is at most 10 per instruction due to the
current expense of running them.  Hopefully, in the future, this can be
increased.

* update use of division instruction

This updates the manner in which the division instruction is used to
confirm to the current requirements (i.e. to provide a witness, a sum
and a witness sum).  These should be relaxed in the near future,
however.

The key purpose of this update is to utilise the fact that the updated
division instruction now supports register splitting and, hence, we can
run this benchmark under KOALABEAR_16.

* correct documentation

A small bits of documentation were not accurately reflecting the truth.

* add missing check for CREATE

There was a missing check that, in the event of a successful creation,
no failure was being claimed.

* disable OOB KOALABEAR_16 check

* fixes for unsafe casts

There were a small number of unsafe casts.

* add in/out indicators on DATA arguments

This adds in / out indicators on arguments to determine whether they
should be considered inputs or outputs.  This helps with auditing.

* fix for JUMPI

CursorBot got me confused, and I updated the JUMPI case incorrectly.